Fixed type error and belongsToInput tests, which previously had a missing `e.preventDefault` issue
