+++ packages/react/src/auto/shadcn/inputs/relationships/ShadcnAutoBelongsToInput.tsx
@@ -68,7 +68,15 @@ export const makeShadcnAutoBelongsToInput = ({
       selectedOption && selectedOption.id ? (
         <Badge key={`selectedRecordTag_${selectedOption.id}`} variant={"outline"}>
           {selectedOption.label}
-          <Button aria-label={`Remove`} onClick={onRemoveRecord} variant="ghost" size="icon">
+          <Button
+            aria-label={`Remove`}
+            onClick={(e) => {
+              e.preventDefault();
+              onRemoveRecord();
+            }}
+            variant="ghost"
+            size="icon"
+          >
             <XIcon />
           </Button>
         </Badge>
@@ -77,7 +85,15 @@ export const makeShadcnAutoBelongsToInput = ({
           <p id={`${danglingSelectedRecordId}`} style={{ color: "red" }}>
             id: {danglingSelectedRecordId}
           </p>
-          <Button aria-label={`Remove`} onClick={onRemoveRecord} variant="ghost" size="icon">
+          <Button
+            aria-label={`Remove`}
+            onClick={(e) => {
+              e.preventDefault();
+              onRemoveRecord();
+            }}
+            variant="ghost"
+            size="icon"
+          >
             <XIcon />
           </Button>
         </Badge>
